requirements-dev.txt
astroid==3.3.9
coverage==7.8.0
crispy-bootstrap4==2025.6
dill==0.3.9
django-shell-plus==1.1.7
django-extensions==4.1
django-crispy-forms==2.4
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.7
pycodestyle==2.13.0
pylint==3.3.6
tomlkit==0.13.2